Refactoring work area presenters inheritance
The whole presenter object hierarchy was creating all kinds of
confusion for me as a maintainer. This is an intermediary step to
provide an even easier refactoring going forward.
