+++ app/views/sipity/controllers/work_areas/core/show.html.curly
